+++ lib/pio/open_flow13.rb
@@ -16,6 +16,7 @@ require 'pio/open_flow13/hello'
 require 'pio/open_flow13/match'
 require 'pio/open_flow13/meter'
 require 'pio/open_flow13/packet_in'
+require 'pio/open_flow13/packet_out'
 require 'pio/open_flow13/send_out_port'
 require 'pio/open_flow13/write_metadata'