+++ lib/pio/open_flow13/flow_mod.rb
@@ -1,5 +1,6 @@
 require 'forwardable'
 require 'pio/open_flow'
+require 'pio/open_flow13/buffer_id'
 require 'pio/open_flow13/match'
 
 # Base module.
@@ -26,22 +27,6 @@ module Pio
 
       def set(value)
         self.command = COMMANDS.fetch(value)
-      end
-    end
-
-    # Buffered packet to apply to, or :no_buffer.
-    class BufferId < BinData::Primitive
-      NO_BUFFER = 0xffffffff
-
-      endian :big
-      uint32 :buffer_id, initial_value: NO_BUFFER
-
-      def get
-        (buffer_id == NO_BUFFER) ? :no_buffer : buffer_id
-      end
-
-      def set(value)
-        self.buffer_id = (value == :no_buffer ? NO_BUFFER : value)
       end
     end