+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 ## develop (unreleased)
 ### New features
+* [#177](https://github.com/trema/pio/pull/177): Add new class `Pio::PacketIn` (OpenFlow1.3).
 * [#173](https://github.com/trema/pio/pull/173): Add support for the OFPXMT\_OFB\_VLAN_{ VID & PCP }.
 * [#174](https://github.com/trema/pio/pull/174): Add support for the OFPXMT\_OFB\_IP\_{ DSCP & ECN }
 * [#178](https://github.com/trema/pio/pull/178): Add support for the OFPXMT\_OFB\_SCTP\_{ SRC & DST }
+
 
 ## 0.21.1 (6/24/2015)
 ### Bugs fixed